# docs/openapi/ferrum.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ferrum",
    "description": "Implementation map of the Ferrum gateway (utoipa dump). Source of truth for GA4GH DRS/WES/TES/TRS/Beacon/htsget is each standard’s published OpenAPI — see docs/GA4GH.md. This file records what this process exposes, including Ferrum-only paths. Not a replacement spec. Not official GA4GH certification. Auth is a deployment choice. Behavioral proof against the published specs is HelixTest.",
    "contact": {
      "name": "Synaptic Four",
      "email": "contact@synapticfour.com"
    },
    "license": {
      "name": "BUSL-1.1",
      "url": "https://github.com/SynapticFour/Ferrum/blob/main/LICENSE"
    },
    "version": "0.3.1"
  },
  "servers": [
    {
      "url": "/",
      "description": "Paths are absolute from the Ferrum gateway origin (default http://127.0.0.1:8080)."
    }
  ],
  "paths": {
...
    "/ga4gh/htsget/v1/reads/{id}": {
      "get": {
        "tags": [
          "handlers"
        ],
        "operationId": "get_reads_ticket",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "DRS object id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "class",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "referenceName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "fields",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "notags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "htsget ticket; urls point at DRS /stream"
          },
          "400": {
            "description": "InvalidInput or UnsupportedFormat (including genomic region params)"
          },
          "404": {
            "description": "NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "handlers"
        ],
        "operationId": "post_reads_ticket",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "DRS object id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "htsget ticket; urls point at DRS /stream"
          },
          "400": {
            "description": "InvalidInput — including POST regions (Ferrum does not slice)"
          }
        }
      }
    },
...
    "/ga4gh/htsget/v1/variants/{id}": {
      "get": {
        "tags": [
          "handlers"
        ],
        "operationId": "get_variants_ticket",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "DRS object id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "class",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "referenceName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "fields",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "notags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "htsget ticket; urls point at DRS /stream"
          },
          "400": {
            "description": "InvalidInput or UnsupportedFormat (including genomic region params)"
          },
          "404": {
            "description": "NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "handlers"
        ],
        "operationId": "post_variants_ticket",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "DRS object id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "htsget ticket; urls point at DRS /stream"
          },
          "400": {
            "description": "InvalidInput — including POST regions (Ferrum does not slice)"
          }
        }
      }
    },
...
  },
...
}